--- a/cmake/glfw.cmake
+++ b/cmake/glfw.cmake
@@ -5,6 +5,8 @@ add_executable(mbgl-glfw
 target_sources(mbgl-glfw
     PRIVATE platform/glfw/glfw_view.hpp
     PRIVATE platform/glfw/glfw_view.cpp
+    PRIVATE platform/glfw/glfw_renderer_frontend.hpp
+    PRIVATE platform/glfw/glfw_renderer_frontend.cpp
     PRIVATE platform/glfw/settings_json.hpp
     PRIVATE platform/glfw/settings_json.cpp
     PRIVATE platform/default/mbgl/util/default_styles.hpp