--- a/cmake/render.cmake
+++ b/cmake/render.cmake
@@ -1,4 +1,6 @@
 add_executable(mbgl-render
+    platform/default/mbgl/renderer/async_renderer_frontend.cpp
+    platform/default/mbgl/renderer/async_renderer_frontend.hpp
     bin/render.cpp
 )