--- a/platform/android/config.cmake
+++ b/platform/android/config.cmake
@@ -57,6 +57,8 @@ macro(mbgl_platform_core)
         PRIVATE platform/default/bidi.cpp
         PRIVATE platform/default/thread_local.cpp
         PRIVATE platform/default/utf.cpp
+        PRIVATE platform/default/mbgl/renderer/default_render_bridge.cpp
+        PRIVATE platform/default/mbgl/renderer/default_render_bridge.hpp
 
         # Image handling
         PRIVATE platform/default/png_writer.cpp