… benchmark/api/query.benchmark.cpp
… benchmark/api/render.benchmark.cpp
… benchmark/src/mbgl/benchmark/util.cpp
… include/mbgl/renderer/query.hpp
… platform/android/src/style/sources/geojson_source.cpp
… platform/android/src/style/sources/geojson_source.hpp
… platform/android/src/style/sources/image_source.cpp
… platform/android/src/style/sources/image_source.hpp
… platform/android/src/style/sources/raster_source.cpp
… platform/android/src/style/sources/raster_source.hpp
… platform/android/src/style/sources/source.cpp
… platform/android/src/style/sources/source.hpp
… platform/android/src/style/sources/sources.cpp
… platform/android/src/style/sources/sources.hpp
… platform/android/src/style/sources/unknown_source.cpp
… platform/android/src/style/sources/unknown_source.hpp
… platform/android/src/style/sources/vector_source.cpp
… platform/android/src/style/sources/vector_source.hpp
… platform/darwin/src/MGLRendererFrontend.h
… platform/darwin/src/MGLShapeSource.mm
… platform/darwin/src/MGLVectorSource.mm
… platform/darwin/src/headless_backend_cgl.cpp
… platform/default/mbgl/gl/headless_backend.cpp
… platform/default/mbgl/gl/headless_backend.hpp
… platform/default/mbgl/gl/headless_display.hpp
… platform/default/mbgl/renderer/async_renderer_frontend.cpp
… platform/default/mbgl/renderer/async_renderer_frontend.hpp
… platform/default/thread_local.cpp
… platform/glfw/glfw_renderer_frontend.cpp
… platform/glfw/glfw_renderer_frontend.hpp