Introduces a `Renderer` interface with a matching `RendererFrontend` interface to decouple the map from the actual rendering. This makes it possible to have different threading strategies per platform.