@ivovandongen 2017-06-21
@jfirebaugh Given your notes in https://github.com/mapbox/mapbox-gl-native/issues/8820 I've started implementing a renderer abstraction and proxy/bridge interface for the SDK to implement so we can do either synchronous or asynchronous rendering. 

As I understood your notes, we want something like:

![render-interface](https://user-images.githubusercontent.com/64527/27399099-88109d74-5670-11e7-9487-3033e5ab1286.png)

Besides quite some details to work out I'm running into a issue with the sdk binding part you might have already thought about:

The proxy/bridge ideally doesn't know about the data passed to the renderer as it contains references to non-public headers (like the TransformState). I can find no way to pass it along with unique_ptrs though, without having the definition been known at the point where I move the data. 
Atm it works when the renderer accepts const references (https://github.com/mapbox/mapbox-gl-native/pull/9327/commits/9f4c3ec1842a301e4fcef08b077abf7640668afe#diff-9f4c7224e9bcf6a1be79b87e1591357dR47). But when using different threads we want to actually move the data and accept a unique_ptr in the renderer as well so that the data is not destroyed before it has been used in the renderer. This invariably leads to `error: invalid application of 'sizeof' to an incomplete type`. A way around this probably would be to use an Impl for each of the messages, but that will add quite a lot of code.