@jfirebaugh 2017-06-21
Current call chain for continuous rendering is:

1. style mutation
2. `onUpdate(Update::Repaint)`
3. `asyncInvalidate.send()`
4. ..........
5. `backend.invalidate()`
6. <SDK code>
7. `Map::render(view)`
8. `renderStyle->update`, `painter->render`

With the introduction of `RenderBridge`, we should reduce the number of moving parts here.

* Replace `onUpdate(Update::Repaint)` with `renderBridge.update(...)`
* Remove `Backend::invalidate`
* Remove `Map::Impl::asyncInvalidate`; instead `RenderBridge` implementation is responsible for throttling repaints
* Merge `View` into `Backend`
* `RenderBridge` must provide `Backend&` to `Renderer::render`
* In `Map` constructor parameters, replace `Backend&` with `RenderBridge&`; remove `createRenderBridge()`

This will make a clean division:

* `RenderBridge` is the interface consumed by `Map` on the main thread.
* `Backend` is the interface consumed by `Renderer` on the orchestration/rendering thread.