@ivovandongen 2017-06-24
@jfirebaugh I made some progress with reducing the number of moving parts. Leaving some notes here.

- [x] Replace onUpdate(Update::Repaint) with renderBridge.update(...)
As discussed, we need to measure if the additional creation of `UpdateParameters` impacts performance. I fixed up the node and mac os bindings. Render tests succeed, but the rendering on the macos app stutters. Need to investigate if this is the cause.

- [x] Remove Backend::invalidate
Removed it from Backend, but kept it intact in the way that a invalidate callback can be provided to the default render bridge so the UI system can be used to synchronise render calls with rendering the rest of the UI like before

- [x] Remove Map::Impl::asyncInvalidate; instead RenderBridge implementation is responsible for coalescing updates/repaints
I've implemented this in the `DefaultRenderBridge` since it is common to a couple of our platforms and I don't want to duplicate the code. ATM it provides either a constructor that takes a callback or the `Backend` and `View` so that `Renderer::render` may be called after coalescing updates

- [x] RenderBridge must provide Backend& to Renderer::render
Done through the invalidate callback / constructor arguments for now. Maybe inheritance is more suitable here. 

- [x] In Map constructor parameters, replace Backend& with RenderBridge&; remove createRenderBridge()
We probably want to split off the MapObserver from the `RenderBridge&` argument so that we more easily use composition in the SDK bindings instead of inheritance. ATM the SDK view implementations are a combination of Backend, RenderBridge, View and MapObserver. Which makes re-use of, `DefaultRenderBridge` for example a bit clunky. For example, in the macos binding: `_mbglView->render(*_mbglView, *_mbglView);`

related: Cutting the ownership between `Map` and `RenderStyle` and providing it indirectly by reference to the `RenderBridge` led me to a bit of an awkward position. As the `RenderStyle` needs to be destructed before the `AnnotationManager` since the former uses the latter through the destructor of `AnnotationTile` we need to ensure that the Renderer/RenderBridge gets destructed before the Map, but since the Map needs the RenderBridge as a reference now we can't declare both on the stack anymore (because we need to explicitly delete them in the same order). 
For example, the "fixed" Annotations test:

```
    std::unique_ptr<RenderBridge> renderBridge;
    std::unique_ptr<Map> map;

    AnnotationTest()
            : renderBridge(std::make_unique<DefaultRenderBridge>(backend, view))
            , map(std::make_unique<Map>(*renderBridge, view.getSize(), 1, fileSource, threadPool, MapMode::Still)) {

    }

    ~AnnotationTest() {
        renderBridge.reset();
        map.reset();
    }
```

When we use the renderer on a separate thread, we are going to have to assure that the destruction of the Renderer/RenderBridge still completes before destructing the map.

- Merge View into Backend
Started on this, but don't know if this makes sense for all implementations, for example `OffscreenTexture` and `OffscreenView`. The View is often re-created as well in the node bindings whereas the Backend and associated Context can remain intact.