@jfirebaugh 2017-06-26
> Removed it from Backend, but kept it intact in the way that a invalidate callback can be provided to the default render bridge so the UI system can be used to synchronise render calls with rendering the rest of the UI like before

I think instead we should have per-SDK implementations of `RenderBackend`, rather than trying to share code with `DefaultRenderBackend`.

> I've implemented this in the `DefaultRenderBridge` since it is common to a couple of our platforms and I don't want to duplicate the code.

It's not that much duplication, and some SDKs probably don't even need it. E.g. `[nativeView setNeedsGLDisplay]` likely already does coalescing internally.

> We probably want to split off the `MapObserver` from the `RenderBridge&` argument

Yes, good idea.

> Cutting the ownership between `Map` and `RenderStyle` and providing it indirectly by reference to the `RenderBridge` led me to a bit of an awkward position.

Does this go away if ownership of `RenderBridge` is passed into `Map`, i.e. the parameter is `std::unique_ptr<RenderBridge>` rather than `RenderBridge&`, and `Map` releases it explicitly in `~Map`?