@ivovandongen 2017-06-26
> I think instead we should have per-SDK implementations of RenderBackend, rather than trying to share code with DefaultRenderBackend.

Makes sense if coallescing is indeed a platform issue. I'll try it out on a couple of platforms. I know that Android also just sets a flag that isn't used until the next draw call.

>> Cutting the ownership between Map and RenderStyle and providing it indirectly by reference to the RenderBridge led me to a bit of an awkward position.

> Does this go away if ownership of RenderBridge is passed into Map, i.e. the parameter is std::unique_ptr<RenderBridge> rather than RenderBridge&, and Map releases it explicitly in ~Map?

That would help a lot. I used a reference initially as moving ownership limits the use of the RenderBridge as a part of the platform View bindings by inheritance (`MGLMapView` for example). But the multiple-inheritance there doesn't make for a very clear abstraction anyway.