@ivovandongen 2017-07-06
I ran the benchmark for the node bindings in `platform/node/test/benchmark.js`

branch | total time | time per request
--- | --- | ---
master | 237046.364ms | 47.41ms
||236751.378ms| 47.35ms
renderer-interface | 401433.224 | 80.28ms
||429934.557ms| 85.99ms

With the following settings:

setting  | value |
--- | --- |
mapPoolSize | 10 |
numRenderings | 5000
failurePercentage | 0
timeoutPercentage | 0
renderingTimeout | 5000
ratio | 2